[dev] refactoring
